backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<5.0
celery>=5.3,<6.0
redis>=5.0,<6.0
drf-spectacular>=0.27,<0.28
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
